requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
django-redis
redis
psycopg2-binary
